Root/Dev_Plan/Phase1/requirements.txt
spacy>=3.5.0
pytest>=7.0.0
markdown>=3.4.0
black>=22.0.0  # for code formatting
flake8>=4.0.0  # for linting
flask>=2.0.0
python-dotenv>=0.19.0
openai>=1.0.0
